# internal/controller/docs/swagger.yaml
definitions:
  models.AddPeopleProjectRequest:
    properties:
      position:
        type: string
      programmer_id:
        type: string
      project_id:
        type: string
    type: object
...
  models.CreateProjectRequest:
    properties:
      attachments:
        type: string
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
    type: object
...
  models.CreateUserRequest:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      position:
        type: string
    type: object
...
  models.EditProjectRequest:
    properties:
      attachments:
        type: string
      end_date:
        type: string
      id:
        type: string
      name:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
...
  models.EditUserRequest:
    properties:
      avatar:
        type: string
      birthday:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      position:
        type: string
      user_id:
        type: string
    type: object
...
  models.GetUserListResponse:
    properties:
      count:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.StandardErrorModel:
    properties:
      error:
        type: string
    type: object
  models.SuccessMessage:
    properties:
      success:
        type: boolean
    type: object
...
info:
  contact:
    name: Soburjon
  description: This is an auto-generated API Docs.
  termsOfService: http://swagger.io/terms/
  title: API
  version: "1"
paths:
  /admin/add-people-project/:
    post:
      consumes:
      - application/json
      description: |-
        team lead tamonidan projectga odam qo'shish
        position ga "team_lead","programmer","intern" larni yozish mumkin
      parameters:
      - description: add_people_project
        in: body
        name: add_people_project
        required: true
        schema:
          $ref: '#/definitions/models.AddPeopleProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: add people project
      tags:
      - admin
  /admin/create-admin/:
    post:
      consumes:
      - application/json
      description: |-
        create admin
        avatar mazil yoziladi
        birthday "2001-02-26" farmatda yoziladi
      parameters:
      - description: create_admin
        in: body
        name: create_admin
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: create admin
      tags:
      - admin
  /admin/create-programmer/:
    post:
      consumes:
      - application/json
      description: |-
        create programmer
        avatar mazil yoziladi
        birthday "2001-02-26" farmatda yoziladi
      parameters:
      - description: create_programmer
        in: body
        name: create_programmer
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: create programmer
      tags:
      - admin
  /admin/create-project/:
    post:
      consumes:
      - application/json
      description: |-
        create project
        attachments mazil yoziladi
        start va end date "2001-02-26" farmatda yoziladi
      parameters:
      - description: create_project
        in: body
        name: create_project
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: create project
      tags:
      - admin
  /admin/delete-admin/{admin_id}/:
    delete:
      consumes:
      - application/json
      description: delete admin
      parameters:
      - description: admin_id
        in: path
        name: admin_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: delete admin
      tags:
      - admin
  /admin/delete-programmer/{programmer_id}/:
    delete:
      consumes:
      - application/json
      description: delete programmer
      parameters:
      - description: programmer_id
        in: path
        name: programmer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: delete programmer
      tags:
      - admin
  /admin/delete-project/{project_id}/:
    delete:
      consumes:
      - application/json
      description: delete project
      parameters:
      - description: project_id
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: delete project
      tags:
      - admin
  /admin/edit-admin/:
    put:
      consumes:
      - application/json
      description: |-
        edit admin
        avatar mazil yoziladi
        birthday "2001-02-26" farmatda yoziladi
      parameters:
      - description: edit_admin
        in: body
        name: edit_admin
        required: true
        schema:
          $ref: '#/definitions/models.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: edit admin
      tags:
      - admin
  /admin/edit-programmer/:
    put:
      consumes:
      - application/json
      description: |-
        edit programmer
        avatar mazil yoziladi
        birthday "2001-02-26" farmatda yoziladi
      parameters:
      - description: edit_programmer
        in: body
        name: edit_programmer
        required: true
        schema:
          $ref: '#/definitions/models.EditUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: edit programmer
      tags:
      - admin
  /admin/edit-project/:
    put:
      consumes:
      - application/json
      description: |-
        edit project
        attachments mazil yoziladi
        start va end date "2001-02-26" farmatda yoziladi
      parameters:
      - description: edit_project
        in: body
        name: edit_project
        required: true
        schema:
          $ref: '#/definitions/models.EditProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: edit project
      tags:
      - admin
  /admin/get-admin-list/:
    get:
      consumes:
      - application/json
      description: get admin list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: get admin list
      tags:
      - admin
  /admin/get-attendance/{programmer_id}/:
    get:
      consumes:
      - application/json
      description: get attendance
      parameters:
      - description: programmer_id
        in: path
        name: programmer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: get attendance
      tags:
      - admin
  /admin/get-my-projects/:
    get:
      consumes:
      - application/json
      description: Admin team lead bolgan projectlari
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetMyProjectsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: get my projects
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"